Functional database
(cherry picked from commit 5c312e725dda80be7bcad2dc3cc1f51904f98342)
