+++ UI2/UI2.csproj
@@ -9,16 +9,15 @@
     </PropertyGroup>
 
     <ItemGroup>
-      <PackageReference Include="Microsoft.EntityFrameworkCore" Version="3.1.10" />
-      <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="3.1.10">
-        <PrivateAssets>all</PrivateAssets>
-        <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-      </PackageReference>
-      <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" Version="3.1.10" />
+      <ProjectReference Include="..\PredictorLibrary\PredictorLibrary.csproj" />
     </ItemGroup>
 
     <ItemGroup>
-      <ProjectReference Include="..\PredictorLibrary\PredictorLibrary.csproj" />
+      <PackageReference Include="Microsoft.EntityFrameworkCore" Version="5.0.0-rc.2.20475.6" />
+      <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="5.0.0-rc.2.20475.6">
+        <PrivateAssets>all</PrivateAssets>
+        <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+      </PackageReference>
     </ItemGroup>
 
 </Project>
